Add "Subject" and "Description" to Templates
Creating complex mail templates in the layout editor is unfeasible; it strips tags and would require a lot of repetition. This allows the template to access the layouts' subjects and descriptions for use outside of the main content block. This is useful for setting a <title> for mail clients that support it, setting hidden preview text for clients that support it (using the description), and creating a title that is isolated from the main content.

While I'm aware you can pass variables to Mail, this eliminates the need to do this for native plugins as well as custom plugins.